Changed pricePerNight from Integer to BigDecimal in UnitPatchDto

+++ src/main/java/com/testtask/booking_system/dto/UnitPatchDto.java
@@ -1,8 +1,9 @@
 package com.testtask.booking_system.dto;
 
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 
 public record UnitPatchDto(
     @Positive(message = "'numberOfRooms' must be greater than 0.") Integer numberOfRooms,
-    @Positive(message = "'pricePerNight' must be greater than 0.") Integer pricePerNight
+    @Positive(message = "'pricePerNight' must be greater than 0.") BigDecimal pricePerNight
 ) {}
